Create SVG map tester.
A test that uses the variant to create a folder of SVGs containing the
map in various states. For example shading all the supply centres to
check that owned supply centres will be visible to the players.

The generated SVGs must be manually inspected for correctness.

Note that this is a WIP - I cannot get the xml encoder to output anything.