Dominance 1939
@zond
There are two // comments in this, I'm not sure if Go can handle that.
The 'default rule' actually overwrites the starting country rule. Let's keep it for now, but in a later stage we might find a solution for this. No need to wait with pushing this for this little feature.

Same for the 'neutral' country - we currently can't colour countries neutral, and that's fine.