Dominance Vietnam
I'm okay with it, but we might want to revisit. It's really a lot of countries so a lot of centres are coloured at the start - but we might not want this amount of colouring.

I'm quite okay with it, but let's see how a couple of matches go and tone it down a bit maybe.