Added orders for Neutral armies
This is a bit of a shot in the dark because I have no clue how to read Go, but I think this is it.

I found an extra two passable variables in SengokuBlank (nil, nil) that I think are both required for Build Anywhere (except for buildanywhere parser) and neutral orders - correct)?